Move `getExpandedCollectionsById` to its own file
